nielsen-excel-automate-feature-historical-pull/pyproject.toml
[project]
name = "excel-nielsen-automate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "appium-python-client~=4.5.1",
    "google-api-python-client~=2.166.0",
    "google-auth-httplib2~=0.2.0",
    "pandas[excel]~=3.0.0",
    "python-dotenv~=1.0.1",
]

[dependency-groups]
dev = [
    "ruff~=0.11.0",
]
